many bugfixes - builds now output .snupkg for whenever i decide to upload to nuget - added disclaimers to synchronizes collections - added SmartMove methods for Monsters, NPCs and Maps - added utility method WithinRangeOfNPCAsync - gameLog and gameError now share a serialized object - fixed reconnect code - IndexedItems are now Indexers - bankPack is now forcefully serialized as lowercase - FindOptimalIndex will now correctly find the first empty slot for stackable items that have no more of themself in the bank - QueuedAction CurrentMS and LengthMS are now floats - rewrote MoveAsync (may need to redo this again due to parallelism) - when being sent to jail, an exception will be thrown with details to help debug why - added missing ponty callback to actually serialize the items - added SellItemAsync - when SmartMoving, if you fail to Town it will automatically regenerate the path without towning enables - when a chest is opened, it will now be correctly removed from the chest collection - the character's instance should now be populated - fixed ICircle.Contains(IPoint) - added more information to CooldownInfo - cyclic buffer should no longer skip every other index - corrected documentation for ChestData.Party - entities should now correctly have their targets updates - rewrote EnumHelper.TryParse for performance reasons - NPC game data should now correctly get enriched - Quantity for all objects will now default to 1 - "p" on Items is now Prediction information - rewrote shortcut finding code for paths with distance specified - priest Heal emit code has been fixed - ShallowMerge can now handle interface inheritance - many other small fixes

+++ AL.SocketClient/Model/QueuedAction.cs
@@ -11,13 +11,13 @@ namespace AL.SocketClient.Model
         ///     The amount of time remaining until the action is completed, in milliseconds.
         /// </summary>
         [JsonProperty("ms")]
-        public int CurrentMS { get; init; }
+        public float CurrentMS { get; init; }
 
         /// <summary>
-        ///     The totel amount of time it takes for the action to complete, in milliseconds.
+        ///     The total amount of time it takes for the action to complete, in milliseconds.
         /// </summary>
         [JsonProperty("len")]
-        public int LengthMS { get; init; }
+        public float LengthMS { get; init; }
 
         /// <summary>
         ///     TODO: Not sure... possibly if multiple actions are being taken at once, this number denotes the order they were
